Addendum patch for bug#21660947
As bzero() of m_complete_data was moved out of init_gci_container
by the patch for this bug, some of the members in the Gci_container
were undefined after the EventBuffer C'tor.
This caused several Valgrind warnings.

This patch introduce a bzero() of m_complete_data in the
EventBuffer C'tor.

A solution with propper C'tor for the Gci_container would have
been better, and is planned for a later patch.