Bug#18560951 --include and --exclude are ignored for fk during restore
The previous fix for this bug(commit 2f5c41cf1c8a) was using
checkDbAndTableName to check if the FK has to be restored or not which
was not right as checkDbAndTableName was checking only table metadata.
This patch reverts that and checks the m_new_tables[] for the child
table inside BackupRestore::fk(). This ensures that the fk gets restored
if and only if the child table gets restored.