Bug#21940542 TOO MUCH SPAM: INNODB: COMPUTE VIRTUAL COLUMN VALUES FAILED
Problem: defining a generated column with an expression of type ROW,
was not prevented and led to a server exit if using the debug binary.

Fix: the expression to fill a generated column must obviously
be of a type which a column can have. MySQL does not support
ROW columns. So, refuse ROW expressions.