Bug#21944866: REFACTORING: REMOVE KILL_BAD_DATA
Follow-up patch: Fix missing error handling in UPDATE
uncovered by the removal of KILL_BAD_DATA.

We did not check if test_quick_select() for processing of WHERE
clauses had failed. This meant that we would continue executing
the UPDATE statement as if nothing had happened. At the end of
the statement an assert could be triggered as we were trying to
send OK to the client but had an error reported.

Before the patch removing KILL_BAD_DATA, this error was masked
if the error during test_quick_select() was an error due to
STRICT mode. That would cause THD::killed to be set to KILL_BAD_DATA
to be set and UPDATE would stop was it seemed the statement
had been killed.

This patch fixes the problem by checking if test_quick_select()
has reported an error and aborting UPDATE if that is the case.