Bug#21810529: CRASH IN ITEM_FUNC::WALK WHEN CODE JUMPS TO GARBAGE LOCATION
When a table is opened for the first time, fix_fields_gcol_func() is called
and it fixes GC expression. Table's memroot is active at that moment.
Item_bool_func2::fix_length_and_dec might call convert_constant_item() to
replace constant expression with evaluated value. When a statement opens a
table with GC, fix_fields is called again and convert_constant_item()
might change the GC's expression again as it doesn't track whether the
constant was converted again. The problem is that the replacement constant
item this time is allocated in the statement's memory and freed when it
ends, while GC lives longer. When next statement comes, it tries to use
freed
item and crashes.

Solution is to make convert_constant_item() to check whether the constant
item is actually needs to be converted - i.e it's not a basic constant or
it's field type is different than field's. Otherwise the conversion isn't
done.