BUG#21508582 UNDO LOG DOES NOT CONTAIN ENOUGH INFORMATION ON SPATIAL COLUMNS
We write undo log for externally stored column as blow:
1. log prefix if there is only prefix index on it;
2. log mbr if there is only spatial index on it;
3. log prefix + mbr if there are both prefix and spatial index on it.

But we don't write enough information in the undo log, we depends on
data dictionary to resolve the undo log. However, the data dictionary
can be changed after the undo log is written.

Now we encode the three cases called spatial status in undo log length,
and keep it compatible with 5.7.9 format.

Reviewed-by: Marko Mäkelä <marko.makela@oracle.com>
Reviewed-by: Jimmy Yang   <jimmy.yang@oracle.com>
RB: 10420