BUG#18518712 - CRASH AFTER TRIGGER READS FROM FREED MEMORY,                WITH ROGUE CHARS IN THE BODY
PROBLEM DESCRIPTION AND FIX:
A trigger that contains NULL or invalid chars upon execution
may result in an ASAN crash. A reparse triggered while executing
the  body of the trigger resulted in a parse error. As part of the
cleanup on parse error, sp_head object is deleted. This results
in freeing the memroot on which thd->lex object is allocated.
Thereafter the thd->lex->sphead is set to NULL. This results in
an invalid memory access.

The fix is not to delete the sp_head on a reparse error (i.e
during execution of trigger/stored routine). The sp_head
object would be deleted at time of drop stored routine/trigger
as is the current behaviour on a successful reparse.