Bug#18554390 : DROP DATABASE NOT WORKING IN PRESENCE OF FOREIGN KEYS
When a Foreign Key is dropped during a node's downtime, the node crashes
when it tries to come up and sync its schema with other nodes during the
start phase. This is because the dropping of FK was never handled in the
Dbdict::restartDropObj().

This patch fixes it by updating that function to handle the dropped FKs
during restart. Also added testcases.