Bug#21929967 GCOLS: GCOL VALUE CHANGES WHEN SESSION CHANGES SQL_MODE
Background:
The value of certain expressions depends on the active sql_mode; the
clearest example is pipes_as_concat: if enabled '||' means CONCAT,
if disabled it means OR.

Problem:
So, if a generated column is created when the mode is off and contains
'||' (with the intention that it means OR), enabling the mode before
selecting the column made the expression change meaning (to CONCAT),
which wasn't the creator's intention.

Views do not have this problem because when they are created, their
definition is written to the FRM file in a "normalized" way: not as
the raw expression typed by the user, but as the result of
Item::print() functions. Those functions print self-sufficient text
("or" or "concat"), which is thus independent of any future change of
sql_mode.

Solution: do as for views: print normalized gcol expression to the
FRM.
Because a gcol expression references only the column's table, the
expression needn't specify the database's and table's names (it would
even make name resolution fail in fix_fields_gcol_func); so we print
with QT_NO_DB | QT_NO_TABLE (QT_NO_DB being the new, clearer name for
QT_COMPACT_FORMAT).
In the process of CREATE TABLE or ALTER TABLE (which write to the
FRM), the raw text is replaced in memory with the normalized text; the
latter must be allocated in the same MEM_ROOT as the former, so it has
the same lifetime. To this aim, MEM_ROOT of the former is remembered.

Tests: the effect is that normalized expressions (written to the FRM,
then read at table open time) are visible in SHOW CREATE TABLE.
A side-effect is that if Item::print() does not print all
characteristics of the Item, the expression in the FRM is incomplete
and can lead to unexpected results. This is the case of WEIGHT_STRING;
as views have the same problem, it can be considered an old issue,
which I have filed as Bug#78783, which will be easy to fix by printing
the right clauses in Item_func_hex_string::print().