Bug #11764714 : Adding --connect-retries and --connect-retry-delay options to NDB tools.
When a tool is not able to connect to the cluster,
it tries reconnecting every 5 seconds by default
(for most tools) for a fixed number of retries.
(12 retries in most tools)
This patch gives the user capability to provide
customized retry delay and number of retries by
giving --connect-retry-delay=X and --connect-retries=Y
options respectively.
(X is the number of seconds after which the tool
tries reconnecting to the cluster and
Y is the number of times the tool will try reconnecting
to the cluster.)

Modified ndb_opts.h file to include --connect-retry-delay
and --connect-retries options as standard options and have
used the values of these options instead of the respective
hard-coded 'retry delay' and 'retries' values in the tools'
source files.