Bug #12630403 	RESET SLAVE SHOULD NOT DELETE ALL ENTRIES FROM MYSQL.NDB_APPLY_STATUS
A new MySQLD option and variable -ndb-clear-apply-status is added.

This new option defaults to ON, which means that the RESET SLAVE command
keeps its current behaviour of deleting all rows in the mysql.ndb_apply_status
table.

When this option is set to OFF, the RESET SLAVE command does not delete any
rows in the mysql.ndb_apply_status table.

This allows users who do not want the default behaviour to have full control
over if / how rows are deleted from mysql.ndb_apply_status.

This is useful for more complex topologies and channel cutover processes.

Arguably, having RESET SLAVE delete all rows exceeds the 'server local' scope
of the command, but this default is preserved for now.

The ndb_rpl_apply_status test case is enhanced to test the new option.

This option can be used to