WL#8438:  Integrate NDB unit testing with MTR and PB2
Problem A: The NDB unit tests located throughout the storage/ndb/ source
tree weren't being picked up by PB2 as it was unable to find the test
executables. This was due to the fact that the executables were
not packed down properly.
Problem B: The unit tests were not being run in MTR, CluB or PB as CTest
could not find the executables. CTest requires the executables to be
added using a specific CMake command which was missing.

Solution for Problem A: Modified the location in which the executables
are built to build/unittests/ndb. This packs down the executables properly.
Solution for Problem B: Created a new macro to add tests. This macro ensures
that the test executables are added properly and thus, can be picked up by
CTest. It also serves to encapsulate the logic for adding a unit test making
the process of building and adding unit tests in the future more convenient.