Bug#21837074 : CLUSTER FAILURE ERROR CODE 4009 IS SHOWN AS 'UNKNOWN ERROR CODE'
When the mysqld is not connected to the cluster, and a query for a ndb table,
is executed, the Ndbapi gives an error 'HA_ERR_NO_CONNECTION'. But it is not
Mapped to a mysqld error code, Hence the output 'UNKNOWN ERROR CODE'.

This Patch resolves the mapping issue of error code 'HA_ERR_NO_CONNECTION'
between the mysql and Ndbapi.
The Patch also includes a testfile and its corresponding result file, to check the fix.