Bug #21950389 SMALL TABLESPACES WITH BLOBS TAKE UP TO 80 TIMES MORE SPACE IN 5.7 THAN IN 5.6
Problem:

1.  For small tablespaces with blobs, the data file size is bigger for 5.7 when
compared with 5.6.  This regression happened because when reservation was done
with FSP_BLOB as the type of the page, the special treatment given to the small
tablespace was avoided.

2.  There was also a problem in the way space reservation was done.
Previously, the total blob size was calculated and the reservation was done in
a single shot.  But, the space reservation done by one mtr is valid only for
that mtr.  Hence this logic also needs to be changed.

Solution:

1. Treat the small tablespaces specially even if the allocation type is
FSP_BLOB.

2.  Reserve the space for each mtr individually.  If for each mtr we are inserting
N blob pages, then it is enough to reserve N pages in the tablespace for each mtr.
There is no need to calculate the total number of blob pages.

rb#10531 approved by Marko.