Patch for bug#21809959
DROPPED EVENT OPERATIONS REMOVED TOO EARLY AFTER INITIAL RESTART

Several objects in the NdbEventOperationImpl is garbage collected
based on which GCI has been consumed by the clients. This also
includes NdbEventOperations which are dropped by dropEventOperation().
However, as the event_op can still have referrences from buffered
events still not consumed, we cant destruct the event_op immediately.
First when all events having a GCI <= the 'event_op->stop_gci' has been
consumed, the event_op can be destructed.

This mechanism relies on a monotonic increasing GCI number - which
it is not! During initial restart the GCI sequence is reset and
starts over again from 0/0. Thus leading to both event_op
being prematurely released, or not released at all (leaked).
See bug report for further explanation.

This patch introduce a MonotonicEpoch, containing both the GCI
and a 'gci generation'. The 'generation' is incremented every
time there is a restart. Comparison operators are provided
as part of the MonotonicEpoch such that the generation is
used to provide a monotonic increasing sequence,