Bug#21927558 ST_INTERSECTION(POLYGON, POLYGON) RETURNS SELF INTERSECTING POLYGON Bug#21927639 ST_SYMDIFFERENCE(POLYGON, POLYGON) RETURNS INVALID MULTIPOLYGOM GEOMETRY Bug#21927733 ST_UNION() PRODUCES AN INVALID POLYGON GEOMETRY Bug#21977775 ST_DIFFERENCE(POLYGON, POLYGON) RETURNS INVALID GEOMETRY
Problem: the problem arises when we have polygons geometries (possibly in multi-polygons)
containing holes such that one of the holes' vertices touches the exterior ring of the
polygon at a point that lies in the interior of a segment of the exterior ring of the
polygon. Suppose now that we have such a polygon and let T be the vertex of a hole of the
polygon that touches the segment AB of the exterior ring (T is a point in the interior of
AB). During a set operation the segment AB may be clipped to another segment CD
containting T, where at least one of C and D lies in the interior of AB. By construction
T must be a touch point that lies in the interior of CD, but numerically (i.e., since C
and D may be computed only approximately) T may be considered as not a point on CD,
which can then indicate that the computed geometry is invalid.

Fix: add preprocessing step in the BG overlay algorithm: before computing the overlay,
compute the self intersection points of each of the two geometries and insert
the points corresponding to any touch interior self intersection that they have.
The solution proposed enforces T to be present on CD after the set operation has been
performed, and ensures the numerical/geometric robustness of the operation.