Bug#22017616: ASSERTION FAILED: TABLE_SHARE->IS_MISSING_PRIMARY_KEY() == M_PREBUILT->CLUST_IND
Problem: Creating a unique index on a virtual point column would
result in assert/error message when trying to access the table. The
root cause was that when checking if a key can be promoted to primary
key, the rejection of virtual columns happened after a special check
for POINT columns which, if successful, would discard further checking
and conclude that the column could be promoted.

Solution: Since virtual columns never can be promoted the check for
this must be precede any specialized checks which disable further
checking.