Bug #22041387: --INITIALIZE DOES CHOWN $DATADIR REGARDLESS,   CAUSING APPARMOR/SELINUX TROUBLE
When --initialize is used with an empty directory owned by the --user
user the server still does chown on the datadir.

This causes SElinux/apparmor to chime in and prevent the otherwise
valid operation that would produce no effect anyway (since the directory
is already owned by the correct user and group id).

Fixed by first checking if the directory is owned by the correct user
and group id before calling chown.