Fixed bug#21927469: VCOL: NOT ABLE TO INSERT INTO (VIRTUAL  NOT NULL) COLUMN
When insert one row with DEFAULT value, the server will do:
* fill_record to update the record[0] with DEFAULT value
* during update, it will check whether the field has a default value
  ** If it has, use the default to update
  ** Otherwise, a warning(an error when sql_mode is STRICT_TRANS_TABLES)
     will be thrown out
For generated columns, they are not allowed to define default value
explicitly(by DEFAULT keyword). The default value of them are calculated
by evaluating the generated expression.

In fill_record, there is a loop which fills each of the values into the
respective field. After this loop, we will update the generated columns
by calling updated_generated_write_fields. However in the first loop,
all of generated columns are involved. This causes the problem of bug
report because the generated expression is not evaluated by far and it
has no defined *default* value.

Solution: In the first loop, we exclude all of the generated columns
because all of their values will be updated after that loop.