Bug #21287796   REGRESSION IN DEALING WITH FILESYSTEM LIMITS
There was loss of precision when calculating the number of
pages in ibdata1. While starting the server we read the file
size from my.cnf and compare it with the size which we get
from os_file_get_size(). When reading from my.cnf we convert
the size into MB (megabyte) and then multiply with number of
pages in 1 MB . The other algorithm calculates the number of
pages by simply dividing the size with page size. These two
methods don't give the same result and hence the server
start fails with size mismatch error.

Algorithm 1
-----------

size=4294967295 ( 4G maximum file size in fat32)
size ( in mb ) = 4294967295 / (1024 * 1024) = 4095
pages in 1 MB (assuming 16K page size) = (1024 *1024)/16384= 64
number of pages =    4095 * 64 = 262080

Algorithm 2
-----------

size = 4294967295
number of pages = 4294967295 / 16384 = 262143

FIX
----
Using the first algorithm to calculate both sizes,so that
results match.

[ Review Done by Annamali and Bin.su  #rb10435 ]