Fixed bug#21779554: CHECK_MISPLACED_ROWS BOGUS "FOUND A  MISPLACED ROW" AND CRASHES
When issue a 'check' command, partition engine will call function
partition_helper::check_misplaced_rows to check whether the rows are in the
right partition. In this function, it will scan the rows of paritions using
rnd_next_in_part function, which misses to update the necessary virtual
generated columns.

To fix this bug, we have to add the base columns of the virtual partitioned
column to the read_set and compute the virtual column values.