BUG#12779463 - windows minidump filename should include the process id and /or timestamp.
For Windows platform original behaviour of mysql server on server crash was
to name minidump file as module_name.dmp where module_name was a name of server
executable file. Such naming scheme doesn't take into account the fact that
every server crash rewrite a minidump left from the previous crash so
improtant information is lost.

To fix this issue new scheme for naming minidump files has been introduced.
Within that new scheme a process identifier is appened to the name of minidump
file. So after this patch a name of the minidump file will look something
like mysqld.exe.7296.dmp, where 7296 is a pid of the mysqld server.