Bug#22062023 - SQL causes MySQL crash
This problem may cause a crash in the following case:

- we have two nested derived tables, denoted idt (the inner one) and
  odt (the outer one), and both derived tables are merged.
- a column c2b in the select list of odt references a column c2a in
  the select list of idt.
- the column c2a in idt contains a scalar subquery.
- the outer query references column c2b of odt, which again references
  c2a of idt.
- the column c2a in odt is not referenced by oq.

This is the query:

SELECT c2b
FROM (SELECT c2a, CASE WHEN c2a <> '' THEN c2a END AS c2b
      FROM (SELECT CASE WHEN c2 IN (SELECT c1 FROM t1) THEN c2 END AS c2a
            FROM t2
           ) AS idt
     ) AS odt;

In this case, SELECT_LEX::delete_unused_merged_columns() will attempt
to delete the subqueries underlying the column odt.c2a that is not
referenced. However, since it references the column idt.c2a,
its subquery will also be deleted, even though this column is actually
used by the outer query, through the reference from c2b.

The solution to the problem is to dive into the expression of the unused
column odt.c2a, checking whether it refers to a used column from an inner
derived table, in which case it is marked as used, which will prevent
deletion of subqueries. A new walker function
Item::propagate_derived_used() was created for this purpose.