BUG#22066787 - MYSQLD DOES NOT SUPPORT PID-FILE,                HANGS SYSTEMD STARTUP.
PIDFile attribute in mysqld unit service file can mismatch with
the pid-file/pid_file attribute specified in my.cnf.

The fix is to specify in pid-file attribute in the unit service
file so that attribute specified in my.cnf is ignored. This shall
prevent in mismatch of attribute between systemd unit file and my.cnf.
User wanting to use a different location, is advised to modify the
unit file.