Bug#21957001 INNODB: FAILING ASSERTION: 0 IN FILE HA_INNOPART.CC LINE 3526
Issue:
======
When one or more/all partitions of a tablespace is discarded, in
ha_innopart::info_low, there is an assertion claiming that the code should
be unreachable, which is the reason for hitting the assertion.

This is handled correctly in ha_innobase::info_low by throwing
appropriate warnings when one or more/all partitions of a tablespace is
discarded.

The same check provided for ha_innobase is added for ha_innopart for the
accumulated total value instead of adding the check for each partition.

Fix:
====
The assertion is removed and appropriate checks and warnings are added for
each partition in ha_innopart::info_low().

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>
RB: 10647