Bug#22117089 RENAME THE TEST INNODB.MONITOR_DEBUG TO INNODB.MONITOR_RESTART
The test innodb.monitor_debug required a debug-instrumented MySQL server
only because it disables the purge.

But, the test has no reason to disable the purge, because there will be
nothing to be purged. The reason is that the test did a slow shutdown
followed by a restart, to ensure that the undo logs are empty.
The test itself is not generating any work for purge, because it is only
executing INSERT, which will not generate any update_undo logs.

Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>