Patch for bug#21441075, updated after review comments
NODE RESTART WITH --INITIAL OPTION BRINGS DOWN THE CLUSTER WITH AN ASSERTION..

Dblqh::execCOPY_FRAGREQ() allocates a ScanRecord used to store the
context of the scan doing the 'COPY_FRAG'. Contrary to a 'normal scan',
the copy-scan does not use Dblqh::initScanrec() to set up the
ScanRecord, but instead does its own init of the ScanRecord members.

Unfortunately the member variable 'prioAFlag' was not initialised,
such that it contained garbage which could later be used to
determine the priority of the copy-scan. Thus, it was rather
random whether a fragment copy executed with normal or increased
A-priority.

Furthermore, in debug mode it could cause an assertion in
Dblqh::next_scanconf_tupkeyreq() when the ScanPtr::prioAFlag
was used as an argument to TupKeyReq::setPrioAFlag():

... 'assert(value <= PRIO_A_MASK)'?