BUG#21915239 SLAVE-PRESERVE-COMMIT-ORDER SHOULD GENERATE ERROR WHEN LOG-BIN IS DISABLED
Problem:
slave-preserve-commit-order only works when the binary log
and log-slave-updates are enabled. However, there was no
error in the slave's error log when slave-preserve-commit-order
is enabled and the binary log or log-slave-updates are
disabled when initializing the slave.

Fix:
Since enabling slave-preserve-commit-order only makes sense
in an MTS scenario, with this patch, an error is printed in
the slave's error log when a multi-threaded slave is initialized
if slave-preserve-commit-order is enabled and either the binary
log or log-slave-updates are disabled. The slave is still able to
initialize with the new error message in the error log but it
will be unable to start because this will also issue an error
(the latter error was already implemented before this patch).
This patch does not add any error on the master server, which is
still able to initialize and start normally even if the slave has
this incompatible configuration.