Bug#20948270: DENIED QUERIES TRUNCATED IN ERROR LOG
The error log subsystem uses a static buffer of
1024 bytes.  The plugin logging system used one
of MYSQL_ERRMSG_SIZE (512 bytes), therefore
unnecessarily cutting messages from plugins to
half their allowed length. This patch changes
code to use MAX_LOG_BUFFER_SIZE (1024) for error
logging, the API for plugins, and for general
logging (to file), so the length will align even
if we change the dimensions later.