Patch for bug#20957068
NDB : EVENT API POLLEVENTS() MUST ACTUALLY CAUSE POLL OF TRANSPORTERS

This patch lets the client thread do its own transporter polling
if it has to wait in pollEvents() / pollEvents2(). Thus we are no
longer dependent of other client threads or the ClusterMgr doing the
poll for us. This removes a problem where pollEvents were slow to
receive events.

Introduction of transporter polling also revealed a problem of
missing mutex protection in ndbcluster_binlog, this has
been added as part of this patch.

The 'enum WaitSignalType' has been extended with the 'WAIT_EVENT'
which is now used as 'tWaitState' when the client is waiting for
completion of a new epoch. NdbImpl::trp_deliver_signal() has been
enhanced to wake up a waiting client in this state whenever a
new epoch is completed.

Implementation of 'the wait' in pollEvent() was moved from
NdbEventBuffer::pollEvent() to Ndb::pollEvent() where also
the transport polling now happends a part of calling
PollGuard::wait_n_unlock()

The above changes made NdbEventBuffer::p_cond obsolete which
is now removed.