Bug#21766380 USER FIELD IN AUDIT LOG IS MANGLED
Issue:
======
USER tag content of the audit log (XML file) was mangled on some platforms:

<USER>????[roo0??&#x4;?^???alhost []</USER>

This was caused by dereferencing a pointer to a char buffer, allocated on a stack,
that was invalid in the current stack context.

Fix:
====
Move allocation of the user name buffer up in the stack position.

Reviewers:
==========
Marcin Babij <marcin.babij@oracle.com>
Arun Kuruvila <arun.kuruvila@oracle.com>