Bug #21318711: WRONG RESULTS FOR TRUNCATED COLUMN AND                AGGREGATION
Issue:
-----
This problem happens under the following conditions:
1) Use implicit grouping.
2) An index with a string column as its first part.
3) WHERE clause has an equality comparision with a string
   that has a few trailing characters in addition to the
   one in the string column's data.

In this case, aggregate function should ideally return a
null. But currently we return a value.

SOLUTION:
---------
In case of equality testing, we need to keep track of
whether saving the constant value went perfectly well. A
truncated value at this stage will generate incorrect
results from an index lookup.

In such a case, don't use the index scan at this stage.