Bug#21789221 SHOW STATUS FAILS WITH LOCK TABLES AND SHOW_COMPATIBILITY_56=OFF
Special case handling of P_S tables inside of LOCK TABLE mode.

Performance Schema tables must be accessible independently of the LOCK TABLE mode.
For example, with SHOW_COMPATIBILITY_56=OFF, SHOW STATUS and SHOW VARIABLES require
implicit access to the corresponding tables in the Performance Schema, regardless
of the LOCK TABLE mode.

Explicit LOCK TABLE is not supported in the Performance Schema, except for the
setup and threads tables.