Bug#22018999: gcols: assertion failed: !error
This problem occurs because we have inconsistent handling of evaluation
of various errors when evaluating constant expressions in generated
columns: Sometimes such errors are reported as errors, sometimes they
are converted to warnings. In strict mode, there is no problem because
they are consistently reported as errors.

The solution to the problem is to temporarily raise strict mode when
checking generated column expressions when creating or altering tables.

Note that this is a change of behavior that should be mentioned in manual.