Bug#22141913 ASSERTION: !"INVALID KEY_PART->FIELDNR" FILL_DD_INDEX_ELEMENTS_FROM_KEY_PARTS
The problem was that the server asserted while creating a altered table.
The reason was that the function dd::fill_dd_columns_from_create_fields()
failed evaluating the columns default values, but we failed to propagate
the error up and continued filling index information which raises a assert
as the column used by index is missing.

A test case is added.